requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
defusedxml==0.7.1
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
httpie==3.2.2
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
psycopg2==2.9.7
Pygments==2.16.1
PyJWT==2.8.0
pyscopg2==66.0.2
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.5.2
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
